--- a/.coveragerc
+++ b/.coveragerc
@@ -3,5 +3,9 @@ omit =
     */python?.?/*
     */site-packages/nose/*
     *__init__*
+exclude_lines =
+    pragma: no cover
+    return NotImplemented
+    raise NotImplementedError
 [xml]
 output = coverage.xml