--- a/.gitignore
+++ b/.gitignore
@@ -28,6 +28,11 @@ pip-log.txt
 .tox
 nosetests.xml
 coverage.xml
+.mypy_cache
+.pytest_cache
+
+# Virtual env
+venv
 
 # Translations
 *.mo
@@ -50,8 +55,6 @@ UpgradeLog.htm
 TestResults/Rx.TE.Tests.mdf
 TestResults/Rx.TE.Tests_log.ldf
 *.user
-
-.mypy_cache
 
 # Cloud9
 .c9