--- a/rx/concurrency/historicalscheduler.py
+++ b/rx/concurrency/historicalscheduler.py
@@ -28,7 +28,7 @@ class HistoricalScheduler(VirtualTimeScheduler):
         """Represents a notion of time for this scheduler. Tasks being scheduled
         on a scheduler will adhere to the time denoted by this property."""
 
-        return self.clock
+        return self._clock
 
     @staticmethod
     def add(absolute, relative):