This supercedes https://github.com/ReactiveX/RxPY/pull/324 -- I've re-considered and decided to put in a little extra effort to submit a somewhat more topical PR.

In particular, I've left out some cosmetic changes which were not directly linked to the PriorityQueue or its related schedulers. They are entirely trivial changes, but polluted the already too large diff because they touch many files.

If we can reach agreement on this PR (I'd be happy to amend in response to suggestions or complaints, of course) I will submit the remaining polish commits at some later time.

As before, I've run every commit separately through Travis, you can see that all of them pass [here](https://github.com/erikkemperman/RxPY/commits/priorityqueue). I believe that the least annoying way to review this is one commit at a time -- nevertheless, I apologize that the PR is still pretty unwieldy...

I also figured I ought to be more verbose in motivating these changes... All links in the following refer to the current master branch, at [b3334408](https://github.com/ReactiveX/RxPY/tree/b333440871ad5e74dbdf5365fe99643f22f92c25):

- I noticed that the PriorityScheduler has some minor, superficial, quirks:
  - It declares a [`capacity`](https://github.com/ReactiveX/RxPY/blob/b333440871ad5e74dbdf5365fe99643f22f92c25/rx/internal/priorityqueue.py#L11) constructor arg, but it is not enforced. I prefer to err on the side of simplicity, so for the moment I propose removing the argument.
  - It declares a [`remove_at`](https://github.com/ReactiveX/RxPY/blob/b333440871ad5e74dbdf5365fe99643f22f92c25/rx/internal/priorityqueue.py#L29) method, but addressing the queue by index isn't a well-defined operation (note that while sorting the list maintains the heap invariant, but conversely a valid heap does not necessarily correspond to a sorted list).
  - The [`peek`](https://github.com/ReactiveX/RxPY/blob/b333440871ad5e74dbdf5365fe99643f22f92c25/rx/internal/priorityqueue.py#L27) method raises a custom InvalidOperationException when the queue is empty, but other methods don't and simply raise IndexError in the same circumstance.
  - It ought to be Generic, in the type of items in the queue.

- The methods of PriorityQueue are synchronized "coarsely", but I notice there is room to be a little more precise with locking the queues, if we define the queue itself to be unsynchronized and shift the burden to the classes that use it:
  - The CurrentThreadScheduler, it seems to me, should not need any explicit locks at all -- its own semantics ought to guarantee safety. But currently it is imposing a [second layer](https://github.com/ReactiveX/RxPY/blob/b333440871ad5e74dbdf5365fe99643f22f92c25/rx/concurrency/currentthreadscheduler.py#L46) of synchronization, on top of the queue's own lock.
  - Also, the CurrentThreadScheduler uses a [dict](https://github.com/ReactiveX/RxPY/blob/b333440871ad5e74dbdf5365fe99643f22f92c25/rx/concurrency/currentthreadscheduler.py#L45) to make sure that each invoking thread is mapped to its own queue. But that is just what `threading.local` does, no need to implement it manually (and leaking keys in the dict, as threads disappear).
  - Refactoring to address the above, we find that there is no longer a need for re-entrant locks and we can use the cheaper regular lock.

- Similar to the preceding arguments, more or less straightforward changes can be made to the other usages of the PriorityQueue, namely the EventLoopScheduler, VirtualTimeScheduler and PyGameScheduler.
  - The EventLoopScheduler, moreover, currently uses a separate [`Timer`](https://github.com/ReactiveX/RxPY/blob/b333440871ad5e74dbdf5365fe99643f22f92c25/rx/concurrency/eventloopscheduler.py#L145) which is unnecessary. Since threads are costly on some platforms, I propose avoiding them when it is easy to do so.
   - When using regular (non re-rentrant) locks, we have to be careful to call the `invoke` methods of scheduled items outside of the synchronized / critical sections -- otherwise recursive scheduling would go amiss. Fortunately the changes needed to achieve this for the affected schedulers aren't too complicated.

By the way, I'm not sure I understand the PyGameScheduler properly: I've made sure that I maintain the semantics that were before, but it seems odd to me that this scheduler doesn't actually rely on any pygame primitives?

Finally, I did include some polishing, but only on the files that were affected already.

Also, to get an idea of the impact on coverage, it might be nice to quickly try what I suggested in issue [#316](https://github.com/ReactiveX/RxPY/issues/316) -- it should take a couple of minutes, and there's no harm in trying since coverage reporting is currently not working?
 
Again, I apologize profusely for the inconvenience of such a large PR -- I hope that you agree there might be some benefit to these changes.