@erikkemperman 2019-03-16
Hmm I just noticed that the little benchmark I included behaves differently on Travis for PRs than regular push builds. That's fixed now, and I've also added a warm-up round to avoid potential bias due to one-time startup artifacts.

Here is the summary:

| Platform | Master | PR     | Link |
|----------|--------|--------|------|
| Linux    | 6.06 s | 5.79 s | [here]  |
| MacOS    | 7.94 s | 7.55 s | [here]  |
| Windows  | 9.64 s | 6.36 s | [here]  |

[1]: https://travis-ci.org/ReactiveX/RxPY/jobs/507137635#L1094
[2]: https://travis-ci.org/ReactiveX/RxPY/jobs/507137637#L991
[3]: https://travis-ci.org/ReactiveX/RxPY/jobs/507137638#L948


I will leave this PR alone now for a while, promise!