@erikkemperman 2019-03-17
Thanks for reviewing, I realize it’s a lot... So the workflow now is it needs one more approval? I see you assigned this to me, does that mean I should merge my own PR if another approval is marked?