@jcafhe 2019-03-17
Your detailed abstract above is much much appreciated :+1: . I like the way you simplify `PriorityQueue` by removing synchronization. 

> - Also, the CurrentThreadScheduler uses a dict to make sure that each invoking thread is mapped to its own queue. But that is just what threading.local does, no need to implement it manually (and leaking keys in the dict, as threads disappear).
> - Refactoring to address the above, we find that there is no longer a need for re-entrant locks and we can use the cheaper regular lock.

I've seen that you have completely remove explicit reference to lock in `CurrentThreadScheduler`. I suppose that everything is done by `threading.local` when en-queuing but I'm not 100% sure, so maybe if you could confirm:
https://github.com/ReactiveX/RxPY/blob/4a924287dc6dc3bb1c7d1199b4b4d9f83776bb89/rx/concurrency/currentthreadscheduler.py#L117-L118

Also the use of  `__new__`  is cool too :+1: