@erikkemperman 2019-03-17
@jcafhe Thanks for taking the time to review! Yeah, the tread local means that each thread will have exclusive access to its own private queue, therefore there is no need for explicit locking — other threads never see the same queue.