@dbrattli 2019-03-17
Cool fix for the CurrentThreadScheduler, and yes with enough approvals you are free to merge yourself.