@erikkemperman 2019-03-17
@dbrattli Thanks. From earlier PR, it seems that the preferred method of merging into master is to squash the commits into a single one? That’s fine by me, especially as there are so many here separated mostly for review purposes — but I was just wondering if it has implications for ability to revert individual changes easily, if needed later?